bpfloader: add option for bpfloader to attach progs to tracepoints
Add support for a new section called "attach", containing a boolean
variable that's true if the programs in the elf file should be
attached to their respective tracepoints.

Also revise the handling of program names in readCodeSections() and
loadCodeSections() to avoid losing information about the tracepoint
associated with a program that's needed to carry out the attachment.

Modify bpfloader to stay alive indefinitely in order to keep programs
attached, and remove its oneshot designation so that it can be
restarted and re-attach programs if it does die.

Finally, update BpfLoadTest to use the new automatic attachment
functionality

Change-Id: I7124a4059786f9b80ad0973a67f862d126fb3432
Signed-off-by: Connor O'Brien <connoro@google.com>